requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiosignal==1.3.2
aiosqlite==0.21.0
alabaster==1.0.0
altair==5.5.0
altgraph==0.17.4
annotated-types==0.7.0
anyio==4.9.0
arrow==1.3.0
arxiv==2.1.3
attrs==25.3.0
babel==2.17.0
beautifulsoup4==4.12.2
bibtexparser==1.4.3
blinker==1.9.0
blis==1.2.0
bs4==0.0.2
cachetools==5.5.2
catalogue==2.0.10
certifi==2025.1.31
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
cloudpathlib==0.21.0
cloudscraper==1.2.71
colorama==0.4.6
coloredlogs==15.0.1
confection==0.1.5
contourpy==1.3.1
Crawl4AI==0.5.0.post8
cryptography==44.0.2
cssselect==1.3.0
cupy-cuda11x==12.3.0
cupy-wheel==12.3.0
cycler==0.12.1
cymem==2.0.11
DataRecorder==3.6.2
Deprecated==1.2.18
distro==1.9.0
docutils==0.21.2
DownloadKit==2.0.7
DrissionPage==4.0.5.6
et_xmlfile==2.0.0
fake-http-header==0.3.5
fake-useragent==2.1.0
fastrlock==0.8.3
faust-cchardet==2.1.19
feedparser==6.0.11
filelock==3.18.0
Flask==3.1.0
fonttools==4.56.0
free_proxy==1.1.3
frozenlist==1.5.0
fsspec==2025.3.2
gitdb==4.0.12
GitPython==3.1.44
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.30.1
humanfriendly==10.0
humanize==4.12.2
idna==3.10
imagesize==1.4.1
impact-factor==1.1.2
importlib_metadata==8.6.1
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.9.0
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kiwisolver==1.4.8
langcodes==3.5.0
language_data==1.3.0
litellm==1.65.1
lxml==5.3.1
marisa-trie==1.2.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.1
matplotlib-venn==1.1.2
mdurl==0.1.2
multidict==6.2.0
murmurhash==1.0.12
narwhals==1.32.0
nest-asyncio==1.6.0
nltk==3.9.1
numpy==1.26.4
openai==1.69.0
openpyxl==3.1.5
outcome==1.3.0.post0
packaging==24.2
pandas==2.2.3
paperscraper==0.2.16
pefile==2023.2.7
pillow==10.4.0
playwright==1.50.0
playwright-stealth==1.0.6
pluggy==1.5.0
preshed==3.0.9
prettytable==3.15.1
propcache==0.3.1
protobuf==5.29.4
psutil==7.0.0
pyarrow==19.0.1
pycparser==2.22
pydantic==2.11.1
pydantic_core==2.33.0
pydeck==0.9.1
pyee==12.1.1
Pygments==2.19.1
pyinstaller==6.12.0
pyinstaller-hooks-contrib==2025.2
pymed_paperscraper==1.0.3
pyOpenSSL==25.0.0
pyparsing==3.2.1
pyperclip==1.9.0
PyQt5==5.15.10
PyQt5-Qt5==5.15.2
PyQt5_sip==12.17.0
pyreadline3==3.5.4
PySocks==1.7.1
pystray==0.19.5
pytest==8.3.5
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
pywin32==306
pywin32-ctypes==0.2.3
PyYAML==6.0.2
rank-bm25==0.2.2
RapidFuzz==3.12.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-file==2.1.0
requests-toolbelt==1.0.0
rich==14.0.0
roman-numerals-py==3.1.0
rpds-py==0.24.0
safetensors==0.5.3
schedule==1.2.1
scholarly==1.7.11
scipy==1.15.2
seaborn==0.13.2
selenium==4.29.0
semanticscholar==0.10.0
sgmllib3k==1.0.0
shellingham==1.5.4
simple-loggers==1.0.5
six==1.17.0
smart-open==7.1.0
smmap==5.0.2
sniffio==1.3.1
snowballstemmer==2.2.0
sortedcontainers==2.4.0
soupsieve==2.6
spacy==3.8.4
spacy-legacy==3.0.12
spacy-loggers==1.0.5
Sphinx==8.2.3
sphinx-rtd-theme==3.0.2
sphinxcontrib-applehelp==2.0.0
sphinxcontrib-devhelp==2.0.0
sphinxcontrib-htmlhelp==2.1.0
sphinxcontrib-jquery==4.1
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==2.0.0
sphinxcontrib-serializinghtml==2.0.0
sql-manager==1.0.5
SQLAlchemy==2.0.39
srsly==2.5.1
streamlit==1.44.0
tenacity==9.0.0
tf-playwright-stealth==1.1.2
thefuzz==0.22.1
thinc==8.3.4
tiktoken==0.9.0
tldextract==5.1.3
tokenizers==0.21.1
toml==0.10.2
tornado==6.4.2
tqdm==4.67.1
transformers==4.50.3
trio==0.29.0
trio-websocket==0.12.2
typer==0.15.2
types-python-dateutil==2.9.0.20241206
typing-inspection==0.4.0
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
wasabi==1.1.3
watchdog==6.0.0
wcwidth==0.2.13
weasel==0.4.1
webdriver-manager==4.0.2
webrequests==1.0.8
websocket-client==1.8.0
Werkzeug==3.1.3
wordcloud==1.9.4
wrapt==1.17.2
wsproto==1.2.0
xxhash==3.5.0
yarl==1.18.3
zipp==3.21.0
